SystemUI: Make QS auto-brightness toggle easier to tap
I often have trouble tapping the auto-brightness toggle with the
tiny 24x24 dp touch target bound by the size of the icon. Material
Design guidelines require touch targets to be at least 48x48 dp
[1], so increase the size of the button to comply. This makes it much
easier to tap, especially on larger devices where the top of the screen
may be harder to reach.

While we're at it, adjust the end margin to be effectively 24 dp. This
isn't consistent with the 16 dp of margin at the start, but it's closer
than before, and 24 dp is more consistent wrt. the start of the button.
More space between the brightness slider and auto-brightness toggle is
also helpful to prevent the ripple from colliding with the end of the
slider.

[1] https://material.io/design/usability/accessibility.html

Change-Id: I66759b23cbac560b549a94e26ac1c8c7e5dc0ac5
Signed-off-by: smokey18 <kmboplays@gmail.com>